Move Reader into the internal module.
