+++ src/Control/Effect/Reader/Internal.hs
@@ -1,2 +1,23 @@
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE ExistentialQuantification #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE StandaloneDeriving #-}
 module Control.Effect.Reader.Internal
-() where
+( Reader(..)
+) where
+
+import Control.Algebra
+
+data Reader r m k
+  = Ask (r -> m k)
+  | forall b . Local (r -> r) (m b) (b -> m k)
+
+deriving instance Functor m => Functor (Reader r m)
+
+instance HFunctor (Reader r) where
+  hmap f (Ask k)       = Ask           (f . k)
+  hmap f (Local g m k) = Local g (f m) (f . k)
+
+instance Effect (Reader r) where
+  handle state handler (Ask k)       = Ask (handler . (<$ state) . k)
+  handle state handler (Local f m k) = Local f (handler (m <$ state)) (handler . fmap k)
